Move configuration generator to new namespace
