+++ src/Util/Configuration/Generator.php
@@ -7,12 +7,12 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace PHPUnit\Util;
+namespace PHPUnit\Util\Configuration;
 
 /**
  * @internal This class is not covered by the backward compatibility promise for PHPUnit
  */
-final class ConfigurationGenerator
+final class Generator
 {
     /**
      * @var string
